refactor(): Add all the free that I can add and correct warning errors

+++ solver/src/err_handling/same_lines_size.c
@@ -14,7 +14,7 @@ int same_lines_size(char *file)
 {
     char *buffer = open_file(file);
     char **arr_lines = convert_str_to_arr(buffer);
-    int size_line = strlen(arr_lines[0]);
+    size_t size_line = strlen(arr_lines[0]);
 
     for (int i = 0; arr_lines[i] != NULL; i++){
         if (strlen(arr_lines[i]) != size_line ||
